Remove inefficient Python server example
